techDebt(buRefactorTraining): Generalise deleteWorker and deleteTraining into deleteRecord in shared utils file
