+++ server/routes/establishments/bulkUpload/validate/shared/utils.js
@@ -1,3 +1,5 @@
 exports.createWorkerKey = (localEstablishmentId, workerId) => {
   return ((localEstablishmentId || '') + (workerId || '')).replace(/\s/g, '');
 };
+
+exports.deleteRecord = (APIRecords, lineNumber) => delete APIRecords[lineNumber];
