+++ server/test/unit/routes/establishments/bulkUpload/validate/shared/utils.spec.js
@@ -1,5 +1,8 @@
 const expect = require('chai').expect;
-const { createWorkerKey } = require('../../../../../../../routes/establishments/bulkUpload/validate/shared/utils');
+const {
+  createWorkerKey,
+  deleteRecord,
+} = require('../../../../../../../routes/establishments/bulkUpload/validate/shared/utils');
 
 describe('utils', () => {
   describe('createWorkerKey', () => {
@@ -25,6 +28,21 @@ describe('utils', () => {
       const workerKey = createWorkerKey(worker.localId, worker.uniqueWorkerId);
 
       expect(workerKey).to.equal('WorkplaceWithSpaces');
+    });
+  });
+
+  describe('deleteRecord', () => {
+    it('should remove worker key/value from myAPIWorkers', async () => {
+      const myAPIWorkers = {
+        2: {},
+        3: {},
+      };
+
+      const workerLineNumber = 3;
+
+      deleteRecord(myAPIWorkers, workerLineNumber);
+
+      expect(!myAPIWorkers[workerLineNumber]).to.equal(true);
     });
   });
 });
