+++ server/test/unit/routes/establishments/bulkUpload/validate/workers/validateDuplicateWorkerID.spec.js
@@ -1,9 +1,7 @@
 const {
   validateDuplicateWorkerID,
 } = require('../../../../../../../routes/establishments/bulkUpload/validate/workers/validateDuplicateWorkerID');
-const {
-  createWorkerKey,
-} = require('../../../../../../../routes/establishments/bulkUpload/validate/workers/validateWorkers');
+const { createWorkerKey } = require('../../../../../../../routes/establishments/bulkUpload/validate/shared/utils');
 const expect = require('chai').expect;
 
 describe('validateDuplicateWorkerID()', () => {
