+++ server/test/unit/routes/establishments/bulkUpload/validate/workers/validateWorkers.spec.js
@@ -1,7 +1,6 @@
 const expect = require('chai').expect;
 const {
   createKeysForWorkers,
-  deleteWorker,
 } = require('../../../../../../../routes/establishments/bulkUpload/validate/workers/validateWorkers');
 
 describe('validateWorkers', () => {
@@ -16,21 +15,6 @@ describe('validateWorkers', () => {
 
       expect(workerKeys[0]).to.equal('AnotherTestWorkplaceUser1');
       expect(workerKeys[1]).to.equal('ThePlaceUser2');
-    });
-  });
-
-  describe('deleteWorker', () => {
-    it('should remove worker key/value from myAPIWorkers', async () => {
-      const myAPIWorkers = {
-        2: {},
-        3: {},
-      };
-
-      const workerLineNumber = 3;
-
-      deleteWorker(myAPIWorkers, workerLineNumber);
-
-      expect(!myAPIWorkers[workerLineNumber]).to.equal(true);
     });
   });
 });
